+++ tests/testthat/test_Logger.R
@@ -3,7 +3,7 @@ context("Logger")
 
 
 
-test_that("active bindings", {
+test_that("set_threshold()", {
   ml <- Logger$new("test_logger")
 
   expect_silent(ml$set_threshold(5))
@@ -13,13 +13,12 @@ test_that("active bindings", {
   expect_error(ml$set_threshold("blubb"), "fatal.*trace")
 
   walk(ml$appenders, function(.x) expect_true(inherits(.x, "Appender")))
-
 })
 
 
 
 
-test_that("basic logging", {
+test_that("$log() & co work", {
   ml <- Logger$new("test_logger", appenders = list(memory = AppenderDt$new()))
   ts <- structure(1540486764.41946, class = c("POSIXct", "POSIXt"))
 
@@ -42,7 +41,7 @@ test_that("basic logging", {
 
 
 
-test_that("logger returns formatted message", {
+test_that("string formatting works", {
   l <- Logger$new("test_logger", propagate = FALSE)
 
   expect_identical(l$fatal("test"), "test")
@@ -53,7 +52,7 @@ test_that("logger returns formatted message", {
 
 
 
-test_that("string formatting is used when appropriate", {
+test_that("string formatting is only used if unnamed arguments exist", {
   lg <- Logger$new(
     "test",
     appenders = list(AppenderConsole$new()),
@@ -72,7 +71,7 @@ test_that("string formatting is used when appropriate", {
 
 
 
-test_that("setting appender threshold works", {
+test_that("set_threshold() works for Loggers and Appenders", {
   lg <- Logger$new("dummy", appenders = AppenderConsole$new())
   lg$set_threshold(200)
   expect_identical(lg$threshold, 200L)
@@ -95,7 +94,7 @@ test_that("setting appender threshold works", {
 
 
 
-test_that("suspending loggers works", {
+test_that("threshold of `0` suspends logging", {
   ml <- Logger$new("test_logger")
 
   expect_output(ml$info("blubb"), "blubb")
@@ -121,7 +120,7 @@ test_that("suspending loggers works", {
 
 
 
-test_that("add/remove appenders", {
+test_that("add_appenders()/remove_appenders()", {
   tf <- tempfile()
   on.exit(unlink(tf))
   ml <- Logger$new("test_logger", appenders = AppenderFile$new(file = tf))
@@ -153,7 +152,7 @@ test_that("add/remove appenders", {
 
 
 
-test_that("modify appenders for a logger", {
+test_that("setting Appender properties works", {
   ml <- Logger$new("test_logger", appenders = list(AppenderConsole$new()), propagate = FALSE)
   tf <- tempfile()
   on.exit(unlink(tf))
@@ -193,7 +192,7 @@ test_that("Exceptions are cought and turned into warnings", {
 
 
 
-test_that("Logger inheritance and event propagation", {
+test_that("Inheritance and event propagation works", {
   tf1 <- tempfile()
   tf2 <- tempfile()
   tf3 <- tempfile()
@@ -222,7 +221,7 @@ test_that("Logger inheritance and event propagation", {
 
 
 
-test_that("thresholds work", {
+test_that("threshold works", {
   c1  <- Logger$new("c1")
   expect_output(c1$error("blubb"), "ERROR")
   c1$set_threshold(100)
@@ -232,16 +231,14 @@ test_that("thresholds work", {
 
 
 
-test_that("ancestry querry works", {
-  tf <- tempfile()
-  on.exit(unlink(tf))
+test_that("$ancestry works", {
+  l4 <- get_logger("l1/l2/l3/l4")
+  l2 <- get_logger("l1/l2")$set_propagate(FALSE)
 
-  l1 <- Logger$new("l1", appenders = AppenderBuffer$new())
-  l2 <- Logger$new("l1/l2", propagate = FALSE, appenders = AppenderConsole$new())
-  l3 <- Logger$new("l1/l2/l3", appenders = AppenderFile$new(tf))
-  l4 <- Logger$new("l1/l2/l3/l4", appenders = AppenderBuffer$new())
-
-  expect_match(format(l4$ancestry), "(->.*){2}.*|")
+  expect_equal(
+    unname(unclass(l4$ancestry)),
+    c(TRUE, FALSE, TRUE, TRUE)
+  )
 })
 
 
@@ -249,7 +246,7 @@ test_that("ancestry querry works", {
 
 # LoggerGlue --------------------------------------------------------------
 
-test_that("LoggerGlue creates custom fields", {
+test_that("LoggerGlue supports custom fields", {
   l <- LoggerGlue$new("glue")
 
   expect_output(l$fatal("test", "test"), glue::glue("test", "test"))
@@ -299,5 +296,6 @@ test_that("$config works with lists", {
   expect_error(l$config(cfg = cfg, list = cfg))
 
   l$config(NULL)
-  expect_true(is_virgin_Logger(l))
+  expect_true(is_virgin_Logger(l, allow_subclass = TRUE))
+  expect_false(is_virgin_Logger(l))
 })