+++ tests/testthat/test_get_logger.R
@@ -3,7 +3,7 @@ context("get_logger")
 
 
 
-test_that("get_logger works as expected", {
+test_that("get_logger(): works as expected", {
 
   lg <- get_logger("blubb")
 
@@ -53,7 +53,7 @@ test_that("get_logger is not confused by existing objects with the same name as 
 
 
 
-test_that("is_virgin_Logger identifies loggers without settings", {
+test_that("is_virgin_Logger() identifies loggers without settings", {
   lg1 <- get_logger("foo/bar")
   expect_true(is_virgin_Logger("foo/bar"))
   lg1$set_threshold("off")
@@ -65,7 +65,7 @@ test_that("is_virgin_Logger identifies loggers without settings", {
 
 
 
-test_that("Creating LoggerGlue with get_logger_glue works as expected", {
+test_that("get_logger_glue() works as expected", {
   lg <- get_logger("log/ger/test")
 
   expect_true(is_Logger(lg))
@@ -78,20 +78,79 @@ test_that("Creating LoggerGlue with get_logger_glue works as expected", {
 
 
 
-test_that("get_logger_glue succeedes to get preconfigured glue loggers", {
+test_that("get_logger_glue() succeedes to get preconfigured glue loggers", {
   lg <- get_logger_glue("log/ger/test")
   lg$set_threshold("fatal")
   lg <- get_logger_glue("log/ger/test")
   expect_s3_class(lg, "LoggerGlue")
-  lg$config(NULL)
+  get_logger("log/ger/test", reset = TRUE)
 })
 
 
 
 
-test_that("get_logger_glue fails to get preconfigured loggers that are not glue loggers", {
+test_that("get_logger_glue() fails to get preconfigured loggers that are not glue loggers", {
   lg <- get_logger("log/ger/test2")
   lg$set_threshold("fatal")
   expect_error(get_logger_glue("log/ger/test2"), "LoggerGlue")
-  lg$config(NULL)
+  get_logger("log/ger/test2", reset = TRUE)
+})
+
+
+
+
+test_that("get_logger(reset == TRUE) completely resets logger", {
+  lg <- get_logger_glue("log/ger/reset")
+  lg$set_threshold("fatal")
+  lg$add_appender(AppenderConsole$new())
+
+  expect_s3_class(lg, "LoggerGlue")
+  expect_identical(lg$threshold, 100L)
+  expect_identical(lg, get_logger("log/ger/reset"))
+
+  lg2 <- get_logger("log/ger/reset", reset = TRUE)
+  lg1 <- get_logger("log/ger/reset")
+  expect_identical(
+    data.table::address(lg1),
+    data.table::address(lg2)
+  )
+
+  lg_g1 <- get_logger_glue("log/ger/reset")
+  lg_g2 <- get_logger_glue("log/ger/reset")
+  expect_identical(lg1, lg2)
+  expect_identical(
+    data.table::address(lg_g1),
+    data.table::address(lg_g2)
+  )
+  expect_false(inherits(lg2, "LoggerGlue"))
+  expect_identical(lg2$threshold, 400L)
+})
+
+
+
+
+test_that("get_logger(reset == TRUE) invalidates old Logger", {
+  lg1 <- get_logger("log/ger/reset", reset = TRUE)
+  lg1$set_threshold("fatal")
+
+  lg2 <- get_logger("log/ger/reset", reset = TRUE)$
+    set_threshold("info")$
+    add_appender(AppenderConsole$new())$
+    set_propagate(FALSE)
+
+  expect_true(!identical(lg1, lg2))
+  expect_warning(lg1$fatal("test"), "log/ger/reset")
+  expect_output(lg2$info("test"))
+
+
+  # invalidation works with logger glue
+  get_logger("log/ger/reset", reset = TRUE)
+  lg1 <- get_logger_glue("log/ger/reset")
+  lg1$set_threshold("fatal")
+
+  lg2 <- get_logger("log/ger/reset", reset = TRUE)
+  expect_true(!identical(lg1, lg2))
+  expect_warning(lg1$fatal("test"), "log/ger/reset")
+  expect_output(lg2$info("test"))
+  get_logger("log/ger/reset", reset = TRUE)
 })