+++ tests/testthat/test_simple_logging.R
@@ -12,7 +12,7 @@ teardown({
 
 
 
-test_that("simple_logging works as expected", {
+test_that("threshold(), console_threshold() and log_exception() work as expected", {
   expect_identical(threshold(), lgr$threshold)
   expect_identical(console_threshold(), lgr$appenders$console$threshold)
 
@@ -55,7 +55,7 @@ test_that("show_log()", {
 
 
 
-test_that("add/remove_appender", {
+test_that("add_appender() and remove_appender() work", {
   tlg <- Logger$new("test", propagate = FALSE)
 
   add_appender(AppenderConsole$new(), target = tlg)
@@ -68,7 +68,7 @@ test_that("add/remove_appender", {
 
 
 
-test_that("option appenders setup", {
+test_that("Option 'lgr.log_file' works", {
   old <- getOption("lgr.log_file")
   on.exit(options("lgr.log_file" = old))
 
@@ -100,4 +100,24 @@ test_that("option appenders setup", {
   options("lgr.log_file" = c(blubb = tempfile('_fail')))
   expect_warning(res <- default_appenders(), "_fail")
   unlink(getOption("lgr.log_file"))
+})
+
+
+
+
+test_that("show_data() works", {
+  basic_config(memory = TRUE)
+  on.exit(basic_config())
+
+  expect_output({
+    lgr$info("a log message")
+    lgr$info("another message with data", data = 1:10)
+    lgr$info(c("a vectorized message", "blubb"))
+  })
+
+  expect_output(show_log())
+  expect_s3_class(show_data(), "data.frame")
+  expect_identical(nrow(show_data()), 4L)
+  expect_s3_class(show_dt(), "data.table")
+  expect_identical(nrow(show_dt()), 4L)
 })