+++ NEWS.md
@@ -1,3 +1,15 @@
+# lgr 0.3.1
+
+  * Added `logger_tree()` which provides an overview of all registered loggers
+  * Added `print()` and `format()` methods for Appenders
+  * `AppenderMemory`: added `data` and `dt` active fields (which return the
+    log as a data.frame or data.table)
+  * Removed deprecated functions `FATAL()`, `ERROR()`. Use `lgr$fatal()`, 
+    `lgr$error()`, ... instead.
+  * `AppenderMemory`: `$buffer_dt()` and `$show()` now handle custom fields
+    containing atomic vectors correctly
+
+
 # lgr 0.3.0
 
   * Added support for rotating log files via `AppenderFileRotating`, 
@@ -8,10 +20,10 @@
   * `AppenderFile$new()` now creates an empty file, or fails if it can't
   * Improved support for RMariaDB and dropped support for RMySQL
   * Improved support for RPostgres and dropped support for RPostgreSQL
-  * removed `FATAL()`, `ERROR()`. Use `lgr$fatal()`, `lgr$error()`, ... instead.
   * added `reset` argument to `get_logger()`. This completely resets the
     configuration of the logger and also replaces special loggers (such as 
     `LoggerGlue`) with vanilla ones.
+
 
 # lgr 0.2.2
 
@@ -29,6 +41,7 @@
 * `with_log_level()` and `with_log_value()` now accept logger names as well as 
   Logger objects as the `logger` argument
 * `get_logger_glue()` now works as intended
+* Deprecated `FATAL()`, `ERROR()`. Use `lgr$fatal()`, `lgr$error()`, ... instead.
 
 
 # lgr 0.2.1